E-mail Support
Hi ,guys. I managed to come up with the mail support function via the
build-in e-mail module. I think it's almost there, but a "CSRF"
exception always appeared in the last step which I really didn't know
how to handle. If we can fix that, we won't bother to wrestle with the
whole mail function in the future. :)